Move logging code to separate file The logging code currently resides in common.py. By moving the code to a separate file, it can be imported independently of the common.py code. It should also enable the use of logging in common.py. The code is moved to keylime_logging.py and includes log_http_response() and init_logging() functions.
The license part is missing in keylime_logging.py and should be filled in.

Resolves #143
