+++ keylime/common.py
@@ -20,7 +20,6 @@ violate any copyrights that exist in this work.
 
 import os.path
 import ConfigParser
-import logging.config
 import sys
 import urlparse
 import json
@@ -181,33 +180,6 @@ def ch_dir(path,logger):
     os.umask(0o077)
     os.chdir(path)
 
-def log_http_response(logger, loglevel, response_body):
-    """Takes JSON response payload and logs error info"""
-    if response_body is None:
-        return False
-    if logger is None:
-        return False
-    
-    log_func = logger.info
-    if loglevel == logging.CRITICAL:
-        log_func = logger.critical
-    elif loglevel == logging.ERROR:
-        log_func = logger.error
-    elif loglevel == logging.WARNING:
-        log_func = logger.warning
-    elif loglevel == logging.INFO:
-        log_func = logger.info
-    elif loglevel == logging.DEBUG:
-        log_func = logger.debug
-    
-    if "results" in response_body and "code" in response_body and "status" in response_body:
-        log_func("Response code %s: %s"%(response_body["code"], response_body["status"]))
-    else:
-        logger.error("Error: unexpected or malformed http response payload")
-        return False
-    
-    return True
-
 def echo_json_response(handler,code,status=None,results=None):
     """Takes a JSON package and returns it to the user w/ full HTTP headers"""
     if handler is None or code is None:
@@ -265,61 +237,6 @@ def get_restful_params(urlstring):
     path_params["api_version"] = api_version
     path_params.update(query_params)
     return path_params
-
-
-LOG_TO_FILE=['cloudagent','registrar','provider_registrar','cloudverifier']
-# not clear that this works right.  console logging may not work
-LOG_TO_SYSCONSOLE=['cloudagent']
-LOG_TO_STREAM=['tenant_webapp']
-LOGDIR='/var/log/keylime'
-if not REQUIRE_ROOT:
-    LOGSTREAM = './keylime-stream.log'
-else:
-    LOGSTREAM=LOGDIR+'/keylime-stream.log'
-
-logging.config.fileConfig(CONFIG_FILE)
-def init_logging(loggername):
-    logger = logging.getLogger("keylime.%s"%(loggername))
-    logging.getLogger("requests").setLevel(logging.WARNING)
-    mainlogger = logging.getLogger("keylime")
-
-    if loggername in LOG_TO_FILE:
-        if not REQUIRE_ROOT:
-            logfilename = "./keylime-all.log"
-        else:
-            logfilename = "%s/%s.log"%(LOGDIR,loggername)
-            if os.getuid()!=0:
-                logger.warning("Unable to log to %s. please run as root"%logfilename)
-                return logger
-            else:
-                if not os.path.exists(LOGDIR):
-                    os.makedirs(LOGDIR, 0o750)
-                chownroot(LOGDIR,logger)
-                os.chmod(LOGDIR,0o750)
-                
-        fh = logging.FileHandler(logfilename)
-        fh.setLevel(logger.getEffectiveLevel())
-        basic_formatter = logging.Formatter('%(created)s - %(name)s - %(levelname)s - %(message)s')
-        fh.setFormatter(basic_formatter)
-        mainlogger.addHandler(fh)
-
-    if loggername in LOG_TO_STREAM:
-        fh = logging.FileHandler(filename=LOGSTREAM,mode='w')
-        fh.setLevel(logger.getEffectiveLevel())
-        basic_formatter = logging.Formatter('%(created)s - %(name)s - %(levelname)s - %(message)s')
-        fh.setFormatter(basic_formatter)
-        mainlogger.addHandler(fh)
-
-    if loggername in LOG_TO_SYSCONSOLE:
-        if os.getuid()!=0:
-            logger.warning("unable to log to /dev/console. please run as root")
-        else:
-            fh = logging.FileHandler("/dev/console")
-            fh.setLevel(logger.getEffectiveLevel())
-            fh.setFormatter(basic_formatter)
-            mainlogger.addHandler(fh)
-    
-    return logger
 
 # this doesn't currently work
 # if LOAD_TEST:
