+++ keylime/keylime_sqlite.py
@@ -19,7 +19,8 @@ violate any copyrights that exist in this work.
 '''
 
 import common
-logger = common.init_logging('keylime_sqlite')
+import keylime_logging
+logger = keylime_logging.init_logging('keylime_sqlite')
 import os
 import sqlite3
 import json
