+++ keylime/provider_registrar.py
@@ -21,7 +21,8 @@ violate any copyrights that exist in this work.
 '''
 
 import common
-logger = common.init_logging('provider-registrar')
+import keylime_logging
+logger = keylime_logging.init_logging('provider-registrar')
 
 import registrar_common
 import ConfigParser
